Migrate To .NET 8
