+++ .gitignore
@@ -107,3 +107,4 @@ Generated_Code #added for RIA/Silverlight projects
 _UpgradeReport_Files/
 Backup*/
 UpgradeLog*.XML
+*/.vs/*
