+++ SMPP/SmppClient/Properties/AssemblyInfo.cs
@@ -1,5 +1,4 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 [assembly: AssemblyTitle("SMPP Client")]
@@ -12,6 +11,3 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
 [assembly: Guid("70e187c1-62c0-4ab2-809b-23f4067ae192")]
-[assembly: AssemblyFileVersion("1.0.13120.1")]
-[assembly: AssemblyInformationalVersion("1.0.13120.1")]
-[assembly: AssemblyVersion("1.0.13120.1")]
