--- SMPP/SmppTestClient/Properties/AssemblyInfo.cs
@@ -1,17 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("SMPP Test Client")]
-[assembly: AssemblyDescription("")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Ardan Studios")]
-[assembly: AssemblyProduct("SMPP Test Client")]
-[assembly: AssemblyCopyright("Copyright © Ardan Studios 2011-2013")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-[assembly: Guid("05a03262-6cf3-4d18-9c99-84946390893e")]
-[assembly: AssemblyFileVersion("1.0.13120.1")]
-[assembly: AssemblyInformationalVersion("1.0.13120.1")]
-[assembly: AssemblyVersion("1.0.13120.1")]
